chore(json-ld): move transformer interfaces and registry to json-ld module
