+++ extensions/common/json-ld/build.gradle.kts
@@ -22,6 +22,7 @@ dependencies {
     api(libs.jakartaJson)
 
     api(project(":spi:common:core-spi"))
+    api(project(":spi:common:transform-spi"))
 
     testImplementation(project(":core:common:junit"))
 }
