+++ extensions/common/json-ld/src/main/java/org/eclipse/edc/jsonld/JsonLdExtension.java
@@ -17,16 +17,20 @@ package org.eclipse.edc.jsonld;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.module.SimpleModule;
 import com.fasterxml.jackson.datatype.jsonp.JSONPModule;
+import org.eclipse.edc.jsonld.transformer.JsonLdTransformerRegistry;
+import org.eclipse.edc.jsonld.transformer.JsonLdTransformerRegistryImpl;
 import org.eclipse.edc.runtime.metamodel.annotation.Extension;
 import org.eclipse.edc.runtime.metamodel.annotation.Inject;
+import org.eclipse.edc.runtime.metamodel.annotation.Provider;
 import org.eclipse.edc.spi.system.ServiceExtension;
 import org.eclipse.edc.spi.system.ServiceExtensionContext;
 import org.eclipse.edc.spi.types.TypeManager;
 
 /**
  * Adds support for working with JSON-LD. Provides an ObjectMapper that works with Jakarta JSON-P
- * types through the TypeManager context {@link #TYPE_MANAGER_CONTEXT_JSON_LD} and provides functions
- * for working with JSON-LD structures.
+ * types through the TypeManager context {@link #TYPE_MANAGER_CONTEXT_JSON_LD} and a registry
+ * for {@link org.eclipse.edc.jsonld.transformer.JsonLdTransformer}s. The module also offers
+ * functions for working with JSON-LD structures.
  */
 @Extension(value = JsonLdExtension.NAME)
 public class JsonLdExtension implements ServiceExtension {
@@ -40,6 +44,11 @@ public class JsonLdExtension implements ServiceExtension {
     @Override
     public String name() {
         return NAME;
+    }
+    
+    @Provider
+    public JsonLdTransformerRegistry jsonLdTransformerRegistry() {
+        return new JsonLdTransformerRegistryImpl();
     }
     
     @Override
